Centraliza funções para recuperar dados doGoogle Spreadsheet em um lugar
