+++ covid19/notifications.py
@@ -1,26 +1,15 @@
-import io
-import requests
-import rows
 from functools import lru_cache
 
 from django.conf import settings
+
+from covid19.google_data import import_info_by_state
 from utils.rocketchat import RocketChat
-
-
-COLLABORATORS_SPREADSHEET_URL = "https://docs.google.com/spreadsheets/d/1S77CvorwQripFZjlWTOZeBhK42rh3u57aRL1XZGhSdI/export?format=csv&id=1S77CvorwQripFZjlWTOZeBhK42rh3u57aRL1XZGhSdI&gid=0"
 
 
 class FakeChat:
 
     def send_message(self, channel, message):
         print(f"New message in {channel}:\n{message}")
-
-
-@lru_cache(maxsize=1)
-def collaborators_data():
-    response = requests.get(COLLABORATORS_SPREADSHEET_URL)
-    url_table = rows.import_from_csv(io.BytesIO(response.content), encoding="utf-8")
-    return url_table
 
 
 @lru_cache(maxsize=1)
@@ -31,11 +20,6 @@ def get_chat():
         chat.user_id = settings.ROCKETCHAT_USER_ID
         chat.auth_token = settings.ROCKETCHAT_AUTH_TOKEN
     return chat
-
-
-def import_info_by_state(state):
-    data = collaborators_data()
-    return [r for r in data if r.uf.upper() == state.upper()][0]
 
 
 def clean_collaborators(collaborators):
@@ -74,8 +58,6 @@ def notify_import_success(spreadsheet):
     chat = get_chat()
     state_info = import_info_by_state(spreadsheet.state)
     channel = "#covid19"
-    collabs = clean_collaborators(state_info.voluntarios)
-    collabs = ' '.join(collabs)
     authors = ' e '.join([
         spreadsheet.user.username,
         spreadsheet.peer_review.user.username,
