+++ covid19/spreadsheet.py
@@ -1,54 +1,12 @@
 import io
-from urllib.parse import parse_qs, urlparse
 
 import rows
-from cachetools import cached, TTLCache
 
 from brazil_data.cities import brazilian_cities_per_state
 from core.util import http_get
+
+from covid19 import google_data
 from covid19.models import StateSpreadsheet
-
-
-STATE_LINKS_SPREADSHEET_ID = "1S77CvorwQripFZjlWTOZeBhK42rh3u57aRL1XZGhSdI"
-FIELDS_BOLETIM = {
-    "date": rows.fields.DateField,
-    "url": rows.fields.TextField,
-    "notes": rows.fields.TextField,
-}
-
-
-def spreadsheet_download_url(url_or_id, file_format):
-    if url_or_id.startswith("http"):
-        spreadsheet_id = parse_qs(urlparse(url_or_id).query)["id"][0]
-    else:
-        spreadsheet_id = url_or_id
-    return f"https://docs.google.com/spreadsheets/u/0/d/{spreadsheet_id}/export?format={file_format}&id={spreadsheet_id}"
-
-
-@cached(cache=TTLCache(maxsize=100, ttl=24 * 3600))
-def get_general_spreadsheet(timeout=5):
-    data = http_get(
-        spreadsheet_download_url(STATE_LINKS_SPREADSHEET_ID, "csv"), timeout
-    )
-    table = rows.import_from_csv(io.BytesIO(data), encoding="utf-8")
-    return {row.uf: row for row in table}
-
-
-def get_state_data_from_google_spreadsheets(state, timeout=5):
-    general_spreadsheet = get_general_spreadsheet()
-    state_spreadsheet_url = general_spreadsheet[state].planilha_brasilio
-    state_spreadsheet_download_url = spreadsheet_download_url(
-        state_spreadsheet_url, "xlsx"
-    )
-    data = http_get(state_spreadsheet_download_url, timeout)
-    reports = rows.import_from_xlsx(
-        io.BytesIO(data), sheet_name="Boletins (FINAL)", force_types=FIELDS_BOLETIM
-    )
-    cases = rows.import_from_xlsx(io.BytesIO(data), sheet_name="Casos (FINAL)")
-    return {
-        "reports": [dict(row._asdict()) for row in reports if row.date],
-        "cases": [dict(row._asdict()) for row in cases if row.municipio]
-    }
 
 
 def get_state_spreadsheets_from_database(state):
@@ -98,7 +56,7 @@ def get_state_data_from_db(state):
 
 def merge_state_data(state):
     # Get data from Google Spreadsheets and DB
-    gs_data = get_state_data_from_google_spreadsheets(state)
+    gs_data = google_data.get_state_data_from_google_spreadsheets(state)
     original_reports = gs_data["reports"]
     original_cases = gs_data["cases"]
     db_data = get_state_data_from_db(state)
@@ -133,6 +91,7 @@ def merge_state_data(state):
         "reports": final_reports,
         "cases": final_cases
     }
+
 
 def create_merged_state_spreadsheet(state):
     state_data = merge_state_data(state)
